requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
groq>=0.4.0
ollama>=0.2.0
requests>=2.31.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
jinja2>=3.1.0
